+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 2.7.0.9
 --------
-* Windows/Linux: The GLUT package compiles without any additional GLUT library now.
+* The GLUT package compiles without any additional library/framework now.
 * Windows: We search for a native freeglut DLL, a MinGW freeglut DLL and a classic GLUT DLL, in that order.
 
 2.7.0.8